+++ web/src/pages/Consumer/locales/zh-CN.ts
@@ -15,16 +15,13 @@
  * limitations under the License.
  */
 export default {
-  'page.consumer.form.itemRuleMessage.username':
-    '最大长度100，仅支持字母、数字和 _ 。',
+  'page.consumer.form.itemRuleMessage.username': '最大长度100，仅支持字母、数字和 _ 。',
   'page.consumer.form.itemExtraMessage.username': '名称需唯一',
-  'page.consumer.notification.warning.enableAuthenticationPlugin': '请至少启用如下一种认证类插件：',
   'page.consumer.username': '名称',
   'page.consumer.username.required': '请输入消费者名称',
   'page.consumer.updateTime': '更新时间',
   'page.consumer.list': '消费者列表',
-  'page.consumer.description':
-    '消费者是路由的消费方，形式包括开发者、最终用户、API 调用等。创建消费者时，需绑定至少一个认证类插件。',
+  'page.consumer.description': '消费者是路由的消费方，形式包括开发者、最终用户、API 调用等。',
   'page.consumer.create': '创建消费者',
   'page.consumer.configure': '配置消费者',
 };