+++ web/src/pages/ServerInfo/List.tsx
@@ -40,7 +40,13 @@ const ServerInfo: React.FC = () => {
         return {
           ...item,
           boot_time: moment(item.boot_time * 1000).format('YYYY-MM-DD HH:mm:ss'),
-          last_report_time: moment(item.last_report_time * 1000).format('YYYY-MM-DD HH:mm:ss'),
+          ...(item.last_report_time
+            ? {
+                last_report_time: moment(item.last_report_time * 1000).format(
+                  'YYYY-MM-DD HH:mm:ss',
+                ),
+              }
+            : {}),
           up_time: moment(item.boot_time * 1000)
             .locale(locale)
             .fromNow(true),